+++ src/ui/src/containers/graphs/styles.module.scss
