+++ src/ui/src/containers/navbar/style.module.scss
