+++ src/ui/src/containers/verticalNavItem/NavItem.tsx
