+++ src/ui/src/containers/verticalNavItem/style.module.scss
