+++ src/ui/src/containers/verticalNavbar/VerticalNavbar.tsx
