+++ src/ui/src/containers/verticalNavbar/style.module.scss
