+++ src/ui/src/pages/jobs/Jobs.tsx
@@ -1,8 +1,5 @@
 import { useLocation } from "react-router";
 
-import VerticalNavbar from "../../components/verticalNavbar/VerticalNavbar";
-import { JobRouteSubpaths } from "../../utils/paths";
-
 const Jobs = () => {
     const { pathname } = useLocation();
     return <div>Current path: {pathname}</div>;
