+++ src/ui/src/utils/constants.ts
@@ -3,3 +3,5 @@ export const ITEMS_PER_PAGE_OPTIONS = [10, 20, 50, 100, 200, 500];
 export const DEFAULT_ITEMS_PER_PAGE = 10;
 
 export const DEFAULT_PAGE = 0;
+
+export const months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
